requirements.txt
adrf==0.1.8
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
Authlib==1.3.2
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.35.51
botocore==1.35.51
cachelib==0.13.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
cryptography==43.0.3
databases==0.9.0
defusedxml==0.7.1
distro==1.9.0
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
drfasyncview==0.2.3
email_validator==2.2.0
et_xmlfile==2.0.0
eventlet==0.37.0
fastapi==0.114.0
feedparser==6.0.11
Flask==3.0.3
flask-oidc==2.2.2
Flask-OpenID==1.3.1
Flask-Session==0.8.0
frozenlist==1.4.1
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
mysqlclient==2.2.6
numpy==2.1.2
openpyxl==3.1.2
packaging==24.1
pandas==2.2.3
passlib==1.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg==3.2.3
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.4.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
s3transfer==0.10.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.38.6
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
yarl==1.15.1
